requirements.txt
# Core dependencies
aiogram>=3.0.0
motor>=3.3.0
aiohttp>=3.8.0
pymongo>=4.5.0

# Environment and configuration
python-dotenv>=1.0.0

# Additional utilities
aiofiles>=23.0.0
asyncio-throttle>=1.0.0

# System monitoring and admin features
psutil>=5.9.0
python-dateutil>=2.8.0

# Backup and logging
# bson>=0.5.10 # bson is part of pymongo, no nned to install separately
